date_trunc raw sql
