+++ sqlalchemy_kusto/dbapi.py
@@ -12,6 +12,120 @@ from azure.kusto.data.exceptions import KustoAuthenticationError, KustoServiceEr
 
 from sqlalchemy_kusto import errors
 
+# PostgreSQL date_trunc grain → T-SQL datepart name
+_GRAIN_TO_TSQL: dict[str, str] = {
+    "microseconds": "millisecond",  # T-SQL has no microsecond
+    "milliseconds": "millisecond",
+    "second": "second",
+    "minute": "minute",
+    "hour": "hour",
+    "day": "day",
+    "week": "week",
+    "week_sun": "week",  # extension: Sunday-based week (not in PostgreSQL)
+    "month": "month",
+    "quarter": "quarter",
+    "year": "year",
+}
+
+# Safe DATEDIFF base epoch per grain.
+# For sub-day grains, epoch 0 (1900-01-01) causes integer overflow; use 2000-01-01.
+# week_sun uses epoch -1 = 1899-12-31 (Sunday), so DATEADD/DATEDIFF stay in Sunday-aligned weeks.
+_GRAIN_EPOCH: dict[str, str] = {
+    "microseconds": "'2000-01-01'",
+    "milliseconds": "'2000-01-01'",
+    "second": "'2000-01-01'",
+    "minute": "'2000-01-01'",
+    "hour": "'2000-01-01'",
+    "day": "0",
+    "week": "0",
+    "week_sun": "-1",
+    "month": "0",
+    "quarter": "0",
+    "year": "0",
+}
+
+
+def _translate_raw_date_trunc(sql: str) -> str:
+    """Translate date_trunc('grain', expr) calls in a raw SQL string to T-SQL DATEADD/DATEDIFF.
+
+    Applied in Cursor.execute() so it covers every execution path including Superset SQLLab,
+    which calls the DBAPI cursor directly via engine.raw_connection(), bypassing dialect hooks.
+    Grain and function name matching are case-insensitive. Unknown grains are left unchanged.
+    """
+    marker = "date_trunc("
+    result: list[str] = []
+    i = 0
+    sql_lower = sql.lower()
+
+    while i < len(sql):
+        pos = sql_lower.find(marker, i)
+        if pos == -1:
+            result.append(sql[i:])
+            break
+
+        result.append(sql[i:pos])
+
+        # Walk forward tracking depth to find the matching ')'.
+        depth = 1
+        j = pos + len(marker)
+        while j < len(sql) and depth > 0:
+            if sql[j] == "(":
+                depth += 1
+            elif sql[j] == ")":
+                depth -= 1
+            j += 1
+
+        if depth != 0:
+            # Unbalanced parentheses — leave the rest unchanged.
+            result.append(sql[pos:])
+            i = len(sql)
+            break
+
+        interior = sql[pos + len(marker) : j - 1]
+
+        # Split interior at the first top-level comma to get grain and expression.
+        comma_pos: int | None = None
+        inner_depth = 0
+        for k, ch in enumerate(interior):
+            if ch == "(":
+                inner_depth += 1
+            elif ch == ")":
+                inner_depth -= 1
+            elif ch == "," and inner_depth == 0:
+                comma_pos = k
+                break
+
+        if comma_pos is None:
+            # Malformed call — leave unchanged.
+            result.append(sql[pos:j])
+            i = j
+            continue
+
+        grain_raw = interior[:comma_pos].strip().strip("'\"")
+        grain = grain_raw.lower()
+        expr = interior[comma_pos + 1 :].strip()
+
+        part = _GRAIN_TO_TSQL.get(grain)
+        if part is None:
+            # Unknown grain — leave the whole call unchanged.
+            result.append(sql[pos:j])
+            i = j
+            continue
+
+        epoch = _GRAIN_EPOCH.get(grain, "0")
+        if grain == "week":
+            # T-SQL DATEDIFF(week,...) counts Sunday boundaries, but epoch 0 is Monday.
+            # Shift the input back 1 day so Sunday stays inside its Mon-Sat week,
+            # producing ISO-standard Monday-based truncation (matches PostgreSQL semantics).
+            result.append(
+                f"DATEADD({part}, DATEDIFF({part}, {epoch}, DATEADD(day, -1, {expr})), {epoch})"
+            )
+        else:
+            result.append(f"DATEADD({part}, DATEDIFF({part}, {epoch}, {expr}), {epoch})")
+        i = j
+
+    return "".join(result)
+
 
 def check_closed(func):
     """Decorator that checks if connection/cursor is closed."""
@@ -203,7 +317,7 @@ class Cursor:
         else:
             self.properties.set_option("query_language", "kql")
 
-        query = Cursor._apply_parameters(operation, parameters)
+        query = Cursor._apply_parameters(_translate_raw_date_trunc(operation), parameters)
         query = query.rstrip()
         try:
             server_response = self.kusto_client.execute(
